feat: custom data fetching for surveys table
