+++ apps/web/app/(app)/environments/[environmentId]/surveys/actions.ts
@@ -1,5 +1,9 @@
 "use server";
 
+import {
+  getSurveyListItem,
+  getSurveyListItems,
+} from "@/app/(app)/environments/[environmentId]/surveys/lib/surveys";
 import { z } from "zod";
 import { authenticatedActionClient } from "@formbricks/lib/actionClient";
 import { checkAuthorization } from "@formbricks/lib/actionClient/utils";
@@ -8,12 +12,7 @@ import {
   getOrganizationIdFromSurveyId,
 } from "@formbricks/lib/organization/utils";
 import { getProducts } from "@formbricks/lib/product/service";
-import {
-  copySurveyToOtherEnvironment,
-  deleteSurvey,
-  getSurvey,
-  getSurveys,
-} from "@formbricks/lib/survey/service";
+import { copySurveyToOtherEnvironment, deleteSurvey } from "@formbricks/lib/survey/service";
 import { generateSurveySingleUseId } from "@formbricks/lib/utils/singleUseSurveys";
 import { ZId } from "@formbricks/types/common";
 import { ZSurveyFilterCriteria } from "@formbricks/types/surveys/types";
@@ -22,7 +21,7 @@ const ZGetSurveyAction = z.object({
   surveyId: ZId,
 });
 
-export const getSurveyAction = authenticatedActionClient
+export const getSurveyListItemAction = authenticatedActionClient
   .schema(ZGetSurveyAction)
   .action(async ({ ctx, parsedInput }) => {
     await checkAuthorization({
@@ -31,7 +30,7 @@ export const getSurveyAction = authenticatedActionClient
       rules: ["survey", "read"],
     });
 
-    return await getSurvey(parsedInput.surveyId);
+    return await getSurveyListItem(parsedInput.surveyId);
   });
 
 const ZCopySurveyToOtherEnvironmentAction = z.object({
@@ -130,7 +129,7 @@ const ZGetSurveysAction = z.object({
   filterCriteria: ZSurveyFilterCriteria.optional(),
 });
 
-export const getSurveysAction = authenticatedActionClient
+export const getSurveyListItemsAction = authenticatedActionClient
   .schema(ZGetSurveysAction)
   .action(async ({ ctx, parsedInput }) => {
     await checkAuthorization({
@@ -141,7 +140,7 @@ export const getSurveysAction = authenticatedActionClient
       rules: ["survey", "read"],
     });
 
-    return await getSurveys(
+    return await getSurveyListItems(
       parsedInput.environmentId,
       parsedInput.limit,
       parsedInput.offset,
