+++ apps/web/app/(app)/environments/[environmentId]/surveys/types/survey.ts
@@ -1,0 +1,26 @@
+import { z } from "zod";
+import { ZSurveyStatus } from "@formbricks/types/surveys/types";
+
+export const ZSurveyListItem = z.object({
+  id: z.string(),
+  name: z.string(),
+  environmentId: z.string(),
+  type: z.enum(["link", "app", "website", "web"]), //we can replace this with ZSurveyType after we remove "web" from schema
+  status: ZSurveyStatus,
+  createdAt: z.date(),
+  updatedAt: z.date(),
+  responseCount: z.number(),
+  creator: z
+    .object({
+      name: z.string(),
+    })
+    .nullable(),
+  singleUse: z
+    .object({
+      enabled: z.boolean(),
+      isEncrypted: z.boolean(),
+    })
+    .nullable(),
+});
+
+export type TSurveyListItem = z.infer<typeof ZSurveyListItem>;
